Updated README
Unfortunately, I didn't have time to pre-record videos of myself walking through demonstrations. I have added what I meant for them to be so that if these slides become a part of the RCOS repo, it'll be clear to those who are taking a look at the slides what is supposed to be going on.